+++ modules/common/app/deadbolt2/common/security/WebIDEHandlerCache.java
@@ -1,9 +1,11 @@
-package deadbolt2.common;
+package deadbolt2.common.security;
 
 import be.objectify.deadbolt.java.ConfigKeys;
 import be.objectify.deadbolt.java.DeadboltHandler;
 import be.objectify.deadbolt.java.ExecutionContextProvider;
 import be.objectify.deadbolt.java.cache.HandlerCache;
+import deadbolt2.common.security.WebIDEDeadboltHandler;
+
 import javax.inject.Inject;
 import javax.inject.Singleton;
 import java.util.HashMap;
